test(carry-over): move the initial_values contract tests to unittests
`test_carry_over_skips_time_independent_variables` and
`test_initial_values_without_time_dim_rejected` build a hand-made in-memory
Study, call `build_problem` once and never reach a solver or a study folder —
the shape of `tests/unittests/gems_runner/simulation/test_integer_strategy.py`,
not of an e2e test. Move them there, with `CONSTANT` defined locally so the
unit tree does not import from the e2e one.

What stays in `tests/e2e/functional` genuinely exercises the pipeline: the
parametrized test drives a full `SimulationSession` over a study folder, and
a `build_problem`-level equivalent would not catch a broken
`effective_carry_over_length`.

Also replace the `carry_over_length` fixture with a `parametrize` decorator
carrying both the config line and the expected pin length, which drops the
None-vs-int ternary from the test body.
